moved cd to xxx_impl.hh to avoid
this way we do not have to add explicit initiation of all
variants of grids and gridView iterators in the library
