+++ flow/flow_blackoil_alugrid.cpp
@@ -16,11 +16,26 @@
   You should have received a copy of the GNU General Public License
   along with OPM.  If not, see <http://www.gnu.org/licenses/>.
 */
-#include "config.h"
-#include <opm/simulators/flow/Main.hpp>
+#include <config.h>
+
 #include <dune/alugrid/grid.hh>
 #include <ebos/eclalugridvanguard.hh>
+#include <opm/simulators/flow/Main.hpp>
+
+// for equilgrid in writer
+// need to include this before eclgenericwriter_impl.hh due to specializations.
+#include <opm/grid/CpGrid.hpp>
+#include <opm/grid/cpgrid/GridHelpers.hpp>
+
+// these are not explicitly instanced in library
+#include <ebos/collecttoiorank_impl.hh>
 #include <ebos/eclgenericproblem_impl.hh>
+#include <ebos/eclgenericthresholdpressure_impl.hh>
+#include <ebos/eclgenerictracermodel_impl.hh>
+#include <ebos/eclgenericwriter_impl.hh>
+#include <ebos/ecltransmissibility_impl.hh>
+#include <ebos/equil/initstateequil_impl.hh>
+
 namespace Opm {
 namespace Properties {
 namespace TTag {
